pyproject.toml
[project]
name = "20250412-invoice-scanner"
version = "0.1.0"
description = "AI-powered invoice parser prototype for the AI x Fintech Build Day hackathon"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "python-multipart>=0.0.6",
    "openai>=1.3.0",
    "pymupdf>=1.23.0",
    "pydantic>=2.4.2",
    "python-dotenv>=1.0.0",
    "pillow>=10.4.0",
    "pytesseract>=0.3.13",
]
